+++ src/plugins/index.ts
@@ -1,2 +1,4 @@
+export * from './gridInitializer';
+export * from './metadataSelector';
 export * from './noDataRenderer';
 export * from './rowSelector';